fix hibernate for ManyToOne and OneToMany and OneToOne annotations

+++ flea-park/src/com/fleaPark/model/products/dao/CategoriaDao.java
@@ -1,3 +1,13 @@
+/*
+ * 
+ * Created by Umberto Ferracci from urania's PC
+ * email: umberto.ferracci@gmail.com
+ * Project: fleaPark
+ * Package: com.fleaPark.model.products.dao
+ * Type: CategoriaDao
+ * Last update: 3-feb-2017 23.55.05
+ * 
+ */
 package com.fleaPark.model.products.dao;
 
 import java.util.ArrayList;
@@ -6,10 +16,9 @@ import java.util.List;
 import com.fleaPark.model.products.Categoria;
 import com.fleaPark.services.persistence.dao.EntityDao;
 import com.fleaPark.services.persistence.dao.EntityDaoHibernate;
-import com.mysql.fabric.xmlrpc.base.Array;
 
-public interface CategoriaDao extends EntityDao<Categoria, Long> {
-    public class CategoriaDaoHibernate extends EntityDaoHibernate<Categoria, Long> implements CategoriaDao {
+public interface CategoriaDao extends EntityDao<Categoria, Integer> {
+    public class CategoriaDaoHibernate extends EntityDaoHibernate<Categoria, Integer> implements CategoriaDao {
 
         @Override
         protected List<Class> dependingClasses() {
